requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
imbalanced-learn==0.12.3
shap==0.46.0
matplotlib==3.9.1
seaborn==0.13.2
pytest==8.3.2
pyyaml==6.0.1
joblib==1.4.2
jupyter==1.0.0
ipykernel==6.29.4